requirements.txt
pandas
regex
requests
selenium 
beautifulsoup4
webdriver-manager
PyYAML
pillow
customtkinter
google-search-results
selenium-stealth